+++ src/Service/CategoryRepository.php
@@ -12,8 +12,8 @@ namespace OxidEsales\GraphQL\Example\Service;
 use OxidEsales\Eshop\Application\Model\Category as CategoryModel;
 use OxidEsales\Eshop\Application\Model\CategoryList as CategoryListModel;
 use OxidEsales\GraphQL\Example\Exception\CategoryNotFound;
-use OxidEsales\GraphQL\Example\DataObject\Category;
-use OxidEsales\GraphQL\Example\DataObject\CategoryFilter;
+use OxidEsales\GraphQL\Example\DataType\Category;
+use OxidEsales\GraphQL\Example\DataType\CategoryFilter;
 
 class CategoryRepository
 {